Split ansatz into different files
