+++ qadence/constructors/__init__.py
@@ -5,7 +5,8 @@ from .feature_maps import (
     exp_fourier_feature_map,
 )
 
-from .ansatze import hea, alt
+from .hea import hea
+from .alt import alt
 
 from .iia import identity_initialized_ansatz
